reverte to b0c2e80 and fix docs
